+++ src/parser/syntax/expr.rs
@@ -5,6 +5,6 @@ pub enum Expr<'a> {
     Add(Box<Expr<'a>>, Box<Expr<'a>>),
     App(Box<Expr<'a>>, Box<Expr<'a>>),
     Ident(&'a str),
-    Nat(i64),
+    Nat(usize),
     Stm(Box<Stm<'a>>)
 }
